Rename package versionbumps to productstreams
