+++ pkg/version/productstreams/product_streams.go
@@ -1,4 +1,4 @@
-package versionbumps
+package productstreams
 
 import (
 	"cmp"
