+++ pkg/version/productstreams/product_streams_test.go
@@ -1,4 +1,4 @@
-package versionbumps
+package productstreams
 
 import (
 	"fmt"
